fixes for invalid drive letter casing for scanfolder
